requirements.txt
openai>=1.50.0
pinecone>=5.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.20.0
dataclasses-json>=0.5.0
